MMCA-4785 | Upgrade Scala & Libraries (#86)
* MMCA-4785 | Update Scala version, http, plugins & dependencies

* MMCA-4785 | Update ErrorHandler signatures and related views

* MMCA-4785 | Update connectors to use HttpClientV2

* MMCA-4785 | Update controllers to pass not_found directly

* MMCA-4785 | Add unapply method for form

* MMCA-4785 | Update repositories & services

* MMCA-4785 | Update views

* MMCA-4785 | Update unit tests

* MMCA-4785 | Cleanup

* MMCA-4785 | Cleanup platops warnings

* MMCA-4785 | sbt-auto-build 3.23.0 incompatible, revert

* MMCA-4785 | Adjust return type, variable naming

+++ app/repositories/RequestedTransactionsCache.scala
@@ -17,6 +17,8 @@
 package repositories
 
 import models.GuaranteeTransactionDates
+import org.mongodb.scala.ToSingleObservablePublisher
+import org.mongodb.scala.SingleObservableFuture
 import org.mongodb.scala.model.Filters.equal
 import org.mongodb.scala.model.Indexes.ascending
 import org.mongodb.scala.model.{IndexModel, IndexOptions, ReplaceOptions}
@@ -39,7 +41,7 @@ class DefaultRequestedTransactionsCache @Inject()(mongoComponent: MongoComponent
         ascending("lastUpdated"),
         IndexOptions()
           .name("requested-transactions-cache-last-updated-index")
-          .expireAfter(config.get[Int]("mongodb.timeToLiveInSeconds"),
+          .expireAfter(config.get[Long]("mongodb.timeToLiveInSeconds"),
             TimeUnit.SECONDS)
       ))) with RequestedTransactionsCache {
